src/test/java/com/blogspot/ostas/leetcode/all/medium/maximum_subarray/SolutionTest.java
package com.blogspot.ostas.leetcode.all.medium.maximum_subarray;

import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;

class SolutionTest {

    /*
      Example 1:
    Input: nums = [-2,1,-3,4,-1,2,1,-5,4]
    Output: 6
    Explanation: The subarray [4,-1,2,1] has the largest sum 6.
    Example 2:
    Input: nums = [1]
    Output: 1
    Explanation: The subarray [1] has the largest sum 1.
    Example 3:
    Input: nums = [5,4,-1,7,8]
    Output: 23
    Explanation: The subarray [5,4,-1,7,8] has the largest sum 23.
      Constraints:
    1 <= nums.length <= 105
    -104 <= nums[i] <= 104
      Follow up: If you have figured out the O(n) solution, try coding another solution using the divide and conquer approach, which is more subtle.
    */
    @Test
    void example_0() {
        var solution = new Solution();
        int[] nums = new int[]{-2, 1, -3, 4, -1, 2, 1, -5, 4};
        int expected = 6;
        var result = solution.maxSubArray(nums);
        assertThat(result).isEqualTo(expected);
    }

    @Test
    void example_1() {
        var solution = new Solution();
        int[] nums = new int[]{1};
        int expected = 1;
        var result = solution.maxSubArray(nums);
        assertThat(result).isEqualTo(expected);
    }

    @Test
    void example_2() {
        var solution = new Solution();
        int[] nums = new int[]{5, 4, -1, 7, 8};
        int expected = 23;
        var result = solution.maxSubArray(nums);
        assertThat(result).isEqualTo(expected);
    }

}
